requirements.txt
selenium==3.141.0
google_auth_oauthlib==0.4.3
google_api_python_client==2.0.2
protobuf==3.15.6
python_dateutil==2.8.1
